+++ src/js/player.js
@@ -1026,6 +1026,12 @@ vjs.Player.prototype.selectSource = function(sources){
     var techName = vjs.capitalize(j[i]),
         tech = window['videojs'][techName];
 
+    // Check if the current tech is defined before continuing
+    if (!tech) {
+      vjs.log.error('The "' + techName + '" tech is undefined. Skipped browser support check for that tech.');
+      continue;
+    }
+
     // Check if the browser supports this technology
     if (tech.isSupported()) {
       // Loop through each source object