+++ CHANGELOG.md
@@ -23,6 +23,7 @@ CHANGELOG
 * Modernized the fullscreen API and added support for IE11 [[view](https://github.com/videojs/video.js/pull/1205)]
 * Added cross-browser testing with SauceLabs, and added Karma as the default test runner ([view](https://github.com/videojs/video.js/pull/1187))
 * Fixed saucelabs integration to run on commits in TravisCI ([view](https://github.com/videojs/video.js/pull/1214))
+* Added a clearer error message when a tech is undefined ([view](https://github.com/videojs/video.js/pull/1210))
 
 --------------------